refactor(ReservationTime): 예약 스케쥴 관련 기능 디렉터리와 엔드포인트 times로 변경
