+++ src/main/java/roomescape/reservationtime/ReservationTimeController.java
@@ -1,8 +1,10 @@
 package roomescape.reservationtime;
 
+import java.time.LocalDate;
 import java.util.List;
 
 import jakarta.validation.Valid;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -11,6 +13,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -44,5 +47,13 @@ public class ReservationTimeController {
     public ResponseEntity<Void> deleteReservationTime(@PathVariable long id) {
         reservationTimeService.deleteReservationTime(id);
         return ResponseEntity.noContent().build();
+    }
+
+    @PostMapping("/{themeId}")
+    public ResponseEntity<ScheduleResponse> getSchedules(@PathVariable long themeId,
+                                                         @RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate date) {
+
+        ScheduleResponse response = reservationTimeService.getSchedules(date, themeId);
+        return ResponseEntity.ok().body(response);
     }
 }
